pyproject.toml
[project]
name = "ai-agents-for-beginners"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "agent-framework-core>=1.0.0b251108",
    "dotenv>=0.9.9",
    "ipdb>=0.13.13",
    "ipython>=9.7.0",
    "openai>=2.7.2",
    "semantic-kernel>=1.36.0",
]
